LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.baselines-expected.txt

2d.text.measure.baselines

Testing baselines


FAIL Testing baselines assert_equals: ctx.measureText('A').ideographicBaseline === 6.25 (got -12.5[number], expected 6.25[number]) expected 6.25 but got -12.5
